+++ src/installNpmPkgs.js
@@ -1,7 +1,8 @@
 // install npm packages
 const childProcess = require('child_process')
 const eslintPkg = require('./version.json')
-module.exports = function ({type}) {
+
+module.exports = function ({ type }) {
     let npmCmd = 'npm i -D'
     for (const key of ['default', type]) {
         const pkgs = eslintPkg[key]
@@ -12,11 +13,10 @@ module.exports = function ({type}) {
     }
     console.log('Installing npm packages...')
     console.log(npmCmd)
-    childProcess.exec(npmCmd, err => {
-        if (err) return console.err('Npm install failed', err)
-        console.log('√ Npm packages installation succeed')
-        setTimeout(() => {
-            console.log('\nPlease restart vscode!')
-        }, 500)
-    })
+    try {
+        childProcess.execSync(npmCmd)
+        console.log('✔ Npm packages installation succeed')
+    } catch (err) {
+        if (err) return console.err('✘ Npm install failed', err)
+    }
 }